update restore for published packages (#43)
* Updating docker test config to newer, less redundant style.

* Specify newest minio version to avoid problems with older versions getting out of sync with aws sdk version we are using.

* Trying out some new make targets.

* Towards getting rid of .env files for tests.

* Remove all .env files. Updates to have all make targets use the same MODULES list.

* Finish off a testing TODO

* Replace deprecated AWS code for tests.

* remove unused map

* update pennsieve-go-core version

* update testify version

* New method to look up files given a package.

* tests for file and folder cases.

* Add WithT method on TestDB.
Update folder_test to use published package.

* working on handle restore folder case.

* Testing

* Fix arg name.

* update logrus version for dependabot

* Fix scoping of testing.T in subtests.

* Make sure TestSourcePackages are not collections.

* Remove debugging Printlns.

* Actually test that we end up with the original S3 version ids when restoring.

* Test parent folder package storage size.

* Apply suggestion from @rohanshah


* Apply suggestion from @rohanshah


* Apply suggestion from @rohanshah


* Apply suggestion from @rohanshah


* Teardown containers when running make test.

---------

+++ lambda/restore/handler/file.go
@@ -62,29 +62,41 @@ func (h *MessageHandler) handleFilePackage(ctx context.Context, orgId int, datas
 				NodeId:       restoreInfo.NodeId})
 		}
 
-		// restore S3 and clean up DynamoDB
-		deleteMarkerResp, err := h.Store.NoSQL.GetDeleteMarkerVersions(ctx, &restoreInfo)
+		// restore S3 if necessary and clean up DynamoDB
+		sourceFiles, err := sqlStore.GetSourceFilesByPackageId(ctx, restoreInfo.Id)
 		if err != nil {
-			return h.errorf("error getting delete record of %s: %w", restoreInfo.NodeId, err)
+			return h.errorf("error looking up source files for package %s: %w", restoreInfo.NodeId, err)
 		}
-		deleteMarker, ok := deleteMarkerResp[restoreInfo.NodeId]
-		if !ok {
-			return h.errorf("no delete record found for %v", restoreInfo)
+		isPublished, err := packageIsPublished(sourceFiles)
+		if err != nil {
+			return h.errorf("publish status error for package %s: %w", restoreInfo.NodeId, err)
 		}
-		sqlStore.LogInfoWithFields(log.Fields{"nodeId": restoreInfo.NodeId, "deleteMarker": *deleteMarker}, "delete marker found")
-		if deleteResponse, err := h.Store.Object.DeleteObjectsVersion(ctx, *deleteMarker); err != nil {
-			return h.errorf("error restoring S3 object %s: %w", *deleteMarker, err)
-		} else if len(deleteResponse.AWSErrors) > 0 {
-			sqlStore.LogErrorWithFields(log.Fields{"nodeId": restoreInfo.NodeId, "s3Info": *deleteMarker}, "AWS error during S3 restore", deleteResponse.AWSErrors)
-			return h.errorf("AWS error restoring S3 object %s: %v", *deleteMarker, deleteResponse.AWSErrors[0])
+		if isPublished {
+			sqlStore.LogInfoWithFields(log.Fields{"nodeId": restoreInfo.NodeId}, "package is published; No S3 delete")
+		} else {
+			deleteMarkerResp, err := h.Store.NoSQL.GetDeleteMarkerVersions(ctx, &restoreInfo)
+			if err != nil {
+				return h.errorf("error getting delete record of %s: %w", restoreInfo.NodeId, err)
+			}
+			deleteMarker, ok := deleteMarkerResp[restoreInfo.NodeId]
+			if !ok {
+				return h.errorf("no delete record found for %v", restoreInfo)
+			}
+			sqlStore.LogInfoWithFields(log.Fields{"nodeId": restoreInfo.NodeId, "deleteMarker": *deleteMarker}, "delete marker found")
+			if deleteResponse, err := h.Store.Object.DeleteObjectsVersion(ctx, *deleteMarker); err != nil {
+				return h.errorf("error restoring S3 object %s: %w", deleteMarker, err)
+			} else if len(deleteResponse.AWSErrors) > 0 {
+				sqlStore.LogErrorWithFields(log.Fields{"nodeId": restoreInfo.NodeId, "s3Info": *deleteMarker}, "AWS error during S3 restore", deleteResponse.AWSErrors)
+				return h.errorf("AWS error restoring S3 object %s: %v", *deleteMarker, deleteResponse.AWSErrors[0])
+			}
 		}
-		if err = h.Store.NoSQL.RemoveDeleteRecords(ctx, []*models.RestorePackageInfo{&restoreInfo}); err != nil {
+		if err = h.Store.NoSQL.RemoveDeleteRecords(ctx, []string{restoreInfo.NodeId}); err != nil {
 			// Don't think this should cause the whole restore to fail
 			sqlStore.LogErrorWithFields(log.Fields{"nodeId": restoreInfo.NodeId, "error": err}, "error removing delete record")
 		}
 
 		// restore dataset storage
-		restoredSize := h.parseSize(deleteMarker)
+		restoredSize := sourceFileSize(sourceFiles)
 		sqlStore.LogInfo("restored size: ", restoredSize)
 		if err = h.restoreStorage(ctx, int64(orgId), datasetId, restoreInfo, restoredSize, sqlStore); err != nil {
 			// Don't think this should fail the whole restore
@@ -192,7 +204,7 @@ func (h *MessageHandler) restoreStorages(ctx context.Context, organizationId, da
 	var totalSize int64
 	sizeByParent := map[int64]int64{}
 	for _, f := range fileInfos {
-		size := h.parseSize(f.S3ObjectInfo)
+		size := sourceFileSize(f.SourceFiles)
 		totalSize += size
 		if f.ParentId != nil {
 			sizeByParent[*f.ParentId] += size
@@ -301,7 +313,7 @@ func (p *NameParts) More() bool {
 
 type RestoreFileInfo struct {
 	*models.RestorePackageInfo
-	*store.S3ObjectInfo
+	SourceFiles []store.File
 }
 
 type RestoreFileInfos []RestoreFileInfo
@@ -314,10 +326,31 @@ func (fs RestoreFileInfos) AsPackageInfos() []*models.RestorePackageInfo {
 	return ps
 }
 
-func (fs RestoreFileInfos) AsS3ObjectInfos() []*store.S3ObjectInfo {
-	os := make([]*store.S3ObjectInfo, len(fs))
+func (fs RestoreFileInfos) PackageNodeIds() []string {
+	os := make([]string, len(fs))
 	for i, f := range fs {
-		os[i] = f.S3ObjectInfo
+		os[i] = f.RestorePackageInfo.NodeId
 	}
 	return os
+}
+
+func packageIsPublished(sourceFiles []store.File) (bool, error) {
+	if len(sourceFiles) == 0 {
+		return false, fmt.Errorf("non-collection package contains no source files")
+	}
+	isPublished := sourceFiles[0].Published
+	for i := 1; i < len(sourceFiles); i++ {
+		if isPublished != sourceFiles[i].Published {
+			return false, fmt.Errorf("package contains a mix of published and unpublished source files")
+		}
+	}
+	return isPublished, nil
+}
+
+func sourceFileSize(sourceFiles []store.File) int64 {
+	var size int64
+	for _, sourceFile := range sourceFiles {
+		size += sourceFile.Size
+	}
+	return size
 }
